requirements.txt
numpy
opencv-python
imutils
pygame
pyyaml